Remove circular dependency to NessieClient in Spark SQL extensions and GC modules
The problem is that the Spark SQL extentions for example rely on iceberg-spark3-runtime, which in turn is compiled against a Nessie version that isn't necessarily the same that we use here. Now when the tests run, we run against a nessie-client version that is different from Iceberg and stuff breaks.

The circular dependecy is fixed by introducing the same NessieClient
version that Iceberg uses.

This also moves NessieClient instantiation into subclasses of
AbstractSparkTest so that each subclass can control how to instantiate
the NessieClient (depending on the Nessie version they pull in)

+++ clients/deltalake/core/src/test/java/org/projectnessie/deltalake/ITDeltaLogBranches.java
@@ -64,17 +64,19 @@ class ITDeltaLogBranches extends AbstractSparkTest {
 
   @AfterEach
   public void closeClient() throws NessieNotFoundException, NessieConflictException {
-    Reference ref = null;
-    try {
-      ref = client.getTreeApi().getReferenceByName("test");
-    } catch (NessieNotFoundException e) {
-      // pass ignore
+    if (null != client) {
+      Reference ref = null;
+      try {
+        ref = client.getTreeApi().getReferenceByName("test");
+      } catch (NessieNotFoundException e) {
+        // pass ignore
+      }
+      if (ref != null) {
+        client.getTreeApi().deleteBranch("test", ref.getHash());
+      }
+      client.close();
+      client = null;
     }
-    if (ref != null) {
-      client.getTreeApi().deleteBranch("test", ref.getHash());
-    }
-    client.close();
-    client = null;
   }
 
   @Test
